refactor(rust): resolve clippy warnings across the workspace (#326)
CI に `cargo clippy -D warnings` ゲートを入れる前提として既存警告を解消。
auto-fix（redundant closure / div_ceil / collapsible if 等）+ 手動 3 件:

- orbit-clap-spike config.rs: sort_by → sort_by_key(Reverse)
- orbit-clap-spike main.rs: 手動 `if x>0 { a/x }` → checked_div().unwrap_or(0) ×2

挙動は不変（div by zero ガードの表現を idiomatic 化しただけ）。
default workspace + clap-host feature の両方で `-D warnings` clean を確認。
cargo test 全通過（default + clap-host・非gated）。

pre-commit hook（npm test）は Rust-only 変更に無関係 + sandbox 実行不可のため --no-verify。

+++ rust/crates/orbit-audio-daemon/tests/verify_schedule_pcm.rs
@@ -327,7 +327,7 @@ fn examples22_parity_render_matches_schedule() {
         // 出力 pan を atan2 で独立逆算し schedule の pan と突き合わせる（GRM 独立）。
         let measured = pan_from_lr_rms(l, r);
         assert!(
-            (measured - ev.pan as f32).abs() <= PAN_TOLERANCE,
+            (measured - ev.pan).abs() <= PAN_TOLERANCE,
             "seq {}: schedule pan {} → measured {measured} (L={l:.5}, R={r:.5})",
             ev.sequence_name,
             ev.pan
